+++ regex/src/com.oracle.truffle.regex/src/com/oracle/truffle/regex/UnsupportedRegexException.java
@@ -40,6 +40,7 @@
  */
 package com.oracle.truffle.regex;
 
+import com.oracle.truffle.api.CompilerDirectives.TruffleBoundary;
 import com.oracle.truffle.api.exception.AbstractTruffleException;
 import com.oracle.truffle.api.interop.ExceptionType;
 import com.oracle.truffle.api.interop.InteropLibrary;
@@ -58,8 +59,9 @@ public final class UnsupportedRegexException extends AbstractTruffleException {
         this.reason = reason;
     }
 
+    @TruffleBoundary
     public UnsupportedRegexException(String reason, Throwable cause) {
-        super(null, cause, UNLIMITED_STACK_TRACE, null);
+        super(cause != null ? cause.getMessage() : null, cause, UNLIMITED_STACK_TRACE, null);
         this.reason = reason;
     }