+++ sdk/src/org.graalvm.polyglot/src/org/graalvm/polyglot/PolyglotException.java
@@ -313,6 +313,16 @@ public final class PolyglotException extends RuntimeException {
     }
 
     /**
+     * Returns <code>true</code> if the application thread was interrupted by an
+     * {@link InterruptedException}.
+     * 
+     * @since 20.3
+     */
+    public boolean isInterrupted() {
+        return impl.isInterrupted();
+    }
+
+    /**
      * Returns <code>true</code> if this exception is caused by an attempt of a guest language
      * program to exit the application using a builtin command. The provided exit code can be
      * accessed using {@link #getExitStatus()}.